+++ boot.py
@@ -301,6 +301,7 @@ import mod3rd
 from mod3rd.admin_esp.wlan import router as router_wlan
 from mod3rd.admin_user.login import router as router_login
 from mod3rd.admin_windup.content import router as router_generators
+from mod3rd.admin_windup.file_api import router as router_file_api
 
 
 logger.info("config done. start windup.")
@@ -313,6 +314,7 @@ if run_not_in_sample_mode:
             router_wlan,
             router_login,
             router_generators,
+            router_file_api,
             status,
             secured_router,
         ] )